Adds support for bulk operations in plugin step validator functionality DynamicsValue/fake-xrm-easy#122
