+++ src/FakeXrmEasy.Plugins/MessageNameConstants.cs
@@ -5,12 +5,14 @@
         internal const string Associate = "Associate";
         internal const string Assign = "Assign";
         internal const string Create = "Create";
+        internal const string CreateMultiple = "CreateMultiple";
         internal const string Delete = "Delete";
         internal const string Disassociate = "Disassociate";
         internal const string Retrieve = "Retrieve";
         internal const string RetrieveMultiple = "RetrieveMultiple";
         internal const string Update = "Update";
+        internal const string UpdateMultiple = "UpdateMultiple";
         internal const string Upsert = "Upsert";
-
+        internal const string UpsertMultiple = "UpsertMultiple";
     }
 }
