--- tests/FakeXrmEasy.Plugins.Tests/PluginSteps/PluginStepValidatorTests.cs
@@ -1,375 +1,0 @@
-﻿using DataverseEntities;
-using FakeXrmEasy.Abstractions.Plugins.Enums;
-using FakeXrmEasy.Plugins.PluginSteps;
-using FakeXrmEasy.Plugins.PluginSteps.InvalidRegistrationExceptions;
-using Xunit;
-
-namespace FakeXrmEasy.Plugins.Tests.PluginSteps
-{
-    public class PluginStepValidatorTests
-    {
-        private readonly PluginStepValidator _validator;
-
-        public PluginStepValidatorTests()
-        {
-            _validator = new PluginStepValidator();
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.AppNotification)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.BackgroundOperation)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.CardStateItem)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.ComponentVersionNrdDataSource)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.ElasticFileAttachment)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.EventExpanderBreadcrumb)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.FlowLog)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.FlowRun)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.MsDynTimelinePin)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.NlsqRegistration)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.None)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.PowerPagesLog)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.RecentlyUsed)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.SearchTelemetry)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.SearchResultsCache)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.SharedWorkspaceAccessToken)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.SharedWorkspaceNr)]
-        public void Should_return_valid_registration_for_upsert_valid_combinations(string messageName, string entityLogicalName)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-            };
-
-            pluginStepDefinition.Stage = ProcessingStepStage.Prevalidation;
-            pluginStepDefinition.Mode = ProcessingStepMode.Synchronous;
-            Assert.True(_validator.IsValid(pluginStepDefinition));
-            
-            pluginStepDefinition.Stage = ProcessingStepStage.Preoperation;
-            pluginStepDefinition.Mode = ProcessingStepMode.Synchronous;
-            Assert.True(_validator.IsValid(pluginStepDefinition));
-            
-            pluginStepDefinition.Stage = ProcessingStepStage.Postoperation;
-            pluginStepDefinition.Mode = ProcessingStepMode.Synchronous;
-            Assert.True(_validator.IsValid(pluginStepDefinition));
-            
-            pluginStepDefinition.Stage = ProcessingStepStage.Postoperation;
-            pluginStepDefinition.Mode = ProcessingStepMode.Asynchronous;
-            Assert.True(_validator.IsValid(pluginStepDefinition));
-        }
-        
-        [Fact]
-        public void Should_return_invalid_registration_for_upsert_and_custom_entities()
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = MessageNameConstants.Upsert,
-                EntityLogicalName = Account.EntityLogicalName,
-            };
-
-            pluginStepDefinition.Stage = ProcessingStepStage.Prevalidation;
-            pluginStepDefinition.Mode = ProcessingStepMode.Synchronous;
-            Assert.Throws<InvalidPrimaryEntityNameException>(() => _validator.IsValid(pluginStepDefinition));
-            
-            pluginStepDefinition.Stage = ProcessingStepStage.Preoperation;
-            pluginStepDefinition.Mode = ProcessingStepMode.Synchronous;
-            Assert.Throws<InvalidPrimaryEntityNameException>(() => _validator.IsValid(pluginStepDefinition));
-            
-            pluginStepDefinition.Stage = ProcessingStepStage.Postoperation;
-            pluginStepDefinition.Mode = ProcessingStepMode.Synchronous;
-            Assert.Throws<InvalidPrimaryEntityNameException>(() => _validator.IsValid(pluginStepDefinition));
-            
-            pluginStepDefinition.Stage = ProcessingStepStage.Postoperation;
-            pluginStepDefinition.Mode = ProcessingStepMode.Asynchronous;
-            Assert.Throws<InvalidPrimaryEntityNameException>(() => _validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.AppNotification, ProcessingStepStage.Prevalidation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.AppNotification, ProcessingStepStage.Preoperation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.SearchTelemetry, ProcessingStepStage.Prevalidation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Upsert, EntityLogicalNameConstants.SearchTelemetry, ProcessingStepStage.Preoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_invalid_for_invalid_registrations_for_upsert_message(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.False(_validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.Upsert, "*", ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Upsert, "*", ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Upsert, "*", ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Upsert, "*", ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_invalid_primary_entity_name_for_upsert(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.Throws<InvalidPrimaryEntityNameException>(() => _validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.Create, "*", ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Create, "*", ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Create, "*", ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Create, "*", ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Create, EntityLogicalNameConstants.Account, ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Create, EntityLogicalNameConstants.Account, ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Create, EntityLogicalNameConstants.Account, ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Create, EntityLogicalNameConstants.Account, ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_valid_registrations_for_create_message_as_valid(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.True(_validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.Create, "*", ProcessingStepStage.Prevalidation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Create, "*", ProcessingStepStage.Preoperation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Create, EntityLogicalNameConstants.Account, ProcessingStepStage.Prevalidation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Create, EntityLogicalNameConstants.Account, ProcessingStepStage.Preoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_invalid_for_invalid_registrations_for_create_message(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.False(_validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.Update, "*", ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Update, "*", ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Update, "*", ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Update, "*", ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Update, EntityLogicalNameConstants.Account, ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Update, EntityLogicalNameConstants.Account, ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Update, EntityLogicalNameConstants.Account, ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Update, EntityLogicalNameConstants.Account, ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_valid_registrations_for_update_message_as_valid(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.True(_validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.Update, "*", ProcessingStepStage.Prevalidation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Update, "*", ProcessingStepStage.Preoperation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Update, EntityLogicalNameConstants.Account, ProcessingStepStage.Prevalidation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Update, EntityLogicalNameConstants.Account, ProcessingStepStage.Preoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_invalid_for_invalid_registrations_for_update_message(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.False(_validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.Retrieve, "*", ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Retrieve, "*", ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Retrieve, "*", ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Retrieve, "*", ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Retrieve, EntityLogicalNameConstants.Account, ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Retrieve, EntityLogicalNameConstants.Account, ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Retrieve, EntityLogicalNameConstants.Account, ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Retrieve, EntityLogicalNameConstants.Account, ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_valid_registrations_for_retrieve_message_as_valid(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.True(_validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.Retrieve, "*", ProcessingStepStage.Prevalidation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Retrieve, "*", ProcessingStepStage.Preoperation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Retrieve, EntityLogicalNameConstants.Account, ProcessingStepStage.Prevalidation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Retrieve, EntityLogicalNameConstants.Account, ProcessingStepStage.Preoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_invalid_for_invalid_registrations_for_retrieve_message(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.False(_validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.RetrieveMultiple, "*", ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.RetrieveMultiple, "*", ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.RetrieveMultiple, "*", ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.RetrieveMultiple, "*", ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.RetrieveMultiple, EntityLogicalNameConstants.Account, ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.RetrieveMultiple, EntityLogicalNameConstants.Account, ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.RetrieveMultiple, EntityLogicalNameConstants.Account, ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.RetrieveMultiple, EntityLogicalNameConstants.Account, ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_valid_registrations_for_retrieve_multiple_message_as_valid(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.True(_validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.RetrieveMultiple, "*", ProcessingStepStage.Prevalidation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.RetrieveMultiple, "*", ProcessingStepStage.Preoperation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.RetrieveMultiple, EntityLogicalNameConstants.Account, ProcessingStepStage.Prevalidation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.RetrieveMultiple, EntityLogicalNameConstants.Account, ProcessingStepStage.Preoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_invalid_for_invalid_registrations_for_retrieve_multiple_message(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.False(_validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.Delete, "*", ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Delete, "*", ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Delete, "*", ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Delete, "*", ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Delete, EntityLogicalNameConstants.Account, ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Delete, EntityLogicalNameConstants.Account, ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Delete, EntityLogicalNameConstants.Account, ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Delete, EntityLogicalNameConstants.Account, ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_valid_registrations_for_delete_message_as_valid(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.True(_validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.Delete, "*", ProcessingStepStage.Prevalidation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Delete, "*", ProcessingStepStage.Preoperation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Delete, EntityLogicalNameConstants.Account, ProcessingStepStage.Prevalidation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Delete, EntityLogicalNameConstants.Account, ProcessingStepStage.Preoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_invalid_for_invalid_registrations_for_delete_message(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.False(_validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.Assign, EntityLogicalNameConstants.Account, ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Assign, EntityLogicalNameConstants.Account, ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Assign, EntityLogicalNameConstants.Account, ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
-        [InlineData(MessageNameConstants.Assign, EntityLogicalNameConstants.Account, ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_valid_registrations_for_any_other_message_by_default(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.True(_validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData(MessageNameConstants.Assign, EntityLogicalNameConstants.Account, ProcessingStepStage.Prevalidation, ProcessingStepMode.Asynchronous)]
-        [InlineData(MessageNameConstants.Assign, EntityLogicalNameConstants.Account, ProcessingStepStage.Preoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_invalid_registrations_for_any_other_message_by_default(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.False(_validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData("*", "*", ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
-        [InlineData("*", "*", ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
-        [InlineData("*", "*", ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
-        [InlineData("*", "*", ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_valid_registrations_for_custom_actions_or_apis_by_default(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.True(_validator.IsValid(pluginStepDefinition));
-        }
-
-        [Theory]
-        [InlineData("*", "*", ProcessingStepStage.Prevalidation, ProcessingStepMode.Asynchronous)]
-        [InlineData("*", "*", ProcessingStepStage.Preoperation, ProcessingStepMode.Asynchronous)]
-        public void Should_return_invalid_for_invalid_registrations_for_custom_actions_or_apis_by_default(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
-        {
-            var pluginStepDefinition = new PluginStepDefinition()
-            {
-                MessageName = messageName,
-                EntityLogicalName = entityLogicalName,
-                Stage = stage,
-                Mode = mode
-            };
-            Assert.False(_validator.IsValid(pluginStepDefinition));
-        }
-    }
-}
